requirements.txt
chromadb==1.0.0
groq==0.22.0
pandas==2.2.3
python-dotenv==1.1.0
semantic_router==0.1.7
streamlit==1.42.0
